changes to file

+++ lib/model/current_weather_model.dart
@@ -111,7 +111,7 @@ class Coord {
 }
 
 class Main {
-  double temp;
+  int temp;
   double feelsLike;
   double tempMin;
   double tempMax;
@@ -132,7 +132,7 @@ class Main {
   });
 
   factory Main.fromJson(Map<String, dynamic> json) => Main(
-        temp: json["temp"].toDouble(),
+        temp: json["temp"].toInt(),
         feelsLike: json["feels_like"].toDouble(),
         tempMin: json["temp_min"].toDouble(),
         tempMax: json["temp_max"].toDouble(),
